Extracts header to its own view
