+++ lib/assets/javascripts/cartodb3/editor/editor-view.js
@@ -4,7 +4,8 @@ var Backbone = require('backbone');
 var AddWidgetsView = require('../components/modals/add-widgets/add-widgets-view');
 var createTextLabelsTabPane = require('../components/tab-pane/create-text-labels-tab-pane');
 var StackLayoutView = require('../components/stack-layout/stack-layout-view');
-var TabPaneTemplate = require('./tab-pane-template.tpl');
+var Header = require('./header.js');
+var TabPaneTemplate = require('./tab-pane.tpl');
 var EditorWidgetsView = require('./widgets/widgets-view');
 
 module.exports = cdb.core.View.extend({
@@ -55,13 +56,16 @@ module.exports = cdb.core.View.extend({
       }
     }];
 
+    var header = new Header({
+      title: this._visDefinitionModel.get('name'),
+      canGoBack: false
+    });
+
     var tabPaneOptions = {
       tabPaneOptions: {
         tagName: 'nav',
         className: 'CDB-NavMenu',
         template: TabPaneTemplate,
-        title: this._visDefinitionModel.get('name'),
-        canGoBack: false,
         tabPaneItemOptions: {
           tagName: 'li',
           className: 'CDB-NavMenu-Item'
@@ -74,6 +78,7 @@ module.exports = cdb.core.View.extend({
     };
 
     this._mapTabPaneView = createTextLabelsTabPane(tabPaneTabs, tabPaneOptions);
+    this.$el.append(header.render().$el);
     this.$el.append(this._mapTabPaneView.render().$el);
 
     // TODO tmp; move to edit-content-view once the top-level tab pane is implemented
