+++ lib/assets/javascripts/cartodb3/editor/header.tpl
@@ -11,7 +11,3 @@
   </div>
 </div>
 
-<nav class="CDB-NavMenu">
-  <ul class="CDB-NavMenu-Inner CDB-Text is-semibold CDB-Size-medium js-menu"></ul>
-  <div class="Editor-content js-content"></div>
-</nav>
