+++ lib/assets/javascripts/cartodb3/editor/tab-pane.tpl
@@ -1,0 +1,4 @@
+<nav class="CDB-NavMenu">
+  <ul class="CDB-NavMenu-Inner CDB-Text is-semibold CDB-Size-medium js-menu"></ul>
+  <div class="Editor-content js-content"></div>
+</nav>
