+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widgets-form-content-view.js
@@ -2,7 +2,8 @@ var cdb = require('cartodb-deep-insights.js');
 var WidgetsFormContentDataView = require('./widgets-form-content-data-view');
 var WidgetsFormContentStyleView = require('./widgets-form-content-style-view');
 var createTextLabelsTabPane = require('../../../components/tab-pane/create-text-labels-tab-pane');
-var TabPaneTemplate = require('../../tab-pane-template.tpl');
+var TabPaneTemplate = require('./tab-pane.tpl');
+var Header = require('../../header.js');
 
 /**
  * View to render all necessary for the widget form
@@ -60,8 +61,6 @@ module.exports = cdb.core.View.extend({
         tagName: 'nav',
         className: 'CDB-NavMenu',
         template: TabPaneTemplate,
-        title: self._widgetDefinitionModel.get('title'),
-        canGoBack: true,
         tabPaneItemOptions: {
           tagName: 'li',
           className: 'CDB-NavMenu-Item'
@@ -73,7 +72,13 @@ module.exports = cdb.core.View.extend({
       }
     };
 
+    var header = new Header({
+      title: this._widgetDefinitionModel.get('title'),
+      canGoBack: true
+    });
+
     this.tabPaneView = createTextLabelsTabPane(this.tabPaneItems, options);
+    this.$el.append(header.render().$el);
     this.$el.append(this.tabPaneView.render().$el);
   },
